Atividade11/requirements.txt
aiohttp>=3.9
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
